requirements.txt
Django==3.0.8
django-environ==0.4.5
djangorestframework==3.11.0
psycopg2==2.8.5
rules==2.2
six==1.14.0